Initial commit of unstable prototype
This script attempts to log into Amazon Vine and then downloads a list of
available products.  Every ten minutes after that, it re-downloads the list
and checks if there are any new entries (it checks the ASINs).

Unfortunately, this script is unstable.  In particular, submitting forms
and getting responses back can result in an HTTP error.  The code can
sometimes retry the operation, but that doesn't always work.

Change-Id: I8ec0564a2f14ca89aa9bbeb9729f8d0ca5eef865
Signed-off-by: Timur Tabi <timur@tabi.org>