Improve stability by reconnecting each time
The Mechanize module allows you to re-use an existing connection to
download more data, but this doesn't appear to work well with Amazon.com.
Or maybe I'm just not using it properly.  Regardless, it's much simpler
and safer to re-connect and re-download the Vine data on each pass.
To achieve that, function get_list() encapsulates the entire process of
logging into Amazon.com and downloading a list of ASINs.

Note that this list appears to be only from the "Your Queue" section, and
I don't know if it supports more than one page yet.  However, it does
appear to be much more stable.

Change-Id: I904349ec5499cca5128700c087daad7f38e23350
Signed-off-by: Timur Tabi <timur@tabi.org>