Abort on general errors
If you have five items already "checked out", Vine won't show you any more
items.  It appears that when this happens, something breaks in the
br.submit() call (I think br somehow becomes None), and we end up with an
infinite General Error loop.  Unfortunately, this is hard to reproduce.

So for now, if we ever get a General Error, just abort.

Change-Id: I9b5ae28e66f48f7a770329c4f52f18cf76228342
Signed-off-by: Timur Tabi <timur@tabi.org>