Prompt for Amazon.com email and password, if needed
Don't rely on environment variables to get the Amazon.com email and password.
The user can now specify these on the command lines, or he will be prompted
to type them in.

Replace the term "userid" with "email", since the Amazon.com userid is really
the email address.  This means that the script now looks for AMAZON_EMAIL
instead of AMAZON_USERID.

Note that if you specify the wrong password, the script will pretend that
nothing is wrong, so that needs to be fixed.

Signed-off-by: Timur Tabi <timur@tabi.org>