Use the tr tag to find ASINs instead of the 'a' tag
Using the 'a' tag to find ASINs is inefficient.  There are multiple 'a'
tags, and using a regex to extract the ASIN is error prone.  Instead, use
the 'tr' tags.  There is one one per ASIN, and the tag itself contains
the ASIN as an attribute.  In addition, the tag contains product category
information in case we ever want that.

Signed-off-by: Timur Tabi <timur@tabi.org>