Go idle on Macs if the screen is off or the user has switched away
We don't want to waste bandwidth checking the Vine website if the user
isn't at the computer.  Currently we only support this feature on Macs.

The script checks two things.  First, it checks if the monitor is on, by
using the 'pmset' program.  This has not been tested on dual-monitor
setups, however.

Second, we check if the user has fast-switched to another account.  We
use a Quartz python module to check this.  It's only been tested on
OS X 10.9.

Signed-off-by: Timur Tabi <timur@tabi.org>