Ignore items that are out of stock
Most of the new Vine For All are, for some reason, not actually in stock
when they appear.  Fortunately, there is a single Javascript variable that
lists all of the items that are out of stock, so just remove those items
from the list.  Now every time the web browser opens a new window for a new
item, you'll actually be able to order it.

Also rename the 'list' variable to 'asins'.

Signed-off-by: Timur Tabi <timur@tabi.org>