Randomize the user-agent
It looks like Amazon is catching on to script.  The Vine web site asks for
a secondary login when the script uses the user-agent that it's been using
for so long.

So now the script uses the fake_useragent module to generate a random
user-agent string every time it logs in.  The script takes about 15 seconds
to initialize its list of candidate strings.