Verify that an item exists before opening a browser window
For some reason, Vine will sometimes list an item on your queue that cannot
actually be ordered.  When you click on the link, you get a 404 (missing
web page) error.

So to avoid wasting the user's time, pre-load the web page to make sure
it actually exists.

Note that this change generates additional Amazon.com logins, which might
annoy Amazon.  I'm working on fixing this.