Add support for account verification
Sometimes Amazon will ask you to verify your account info because it thinks
that someone is trying to hack into your account.  When this happens,
prompt the user to enter the information asked by the web site (only the
first question is supported), and then submit the form.

Unfortunately, this code has not been tested.  If this prompt ever
appears for you, please email me the response.html file.