Login only once per pass, and support image captchas
Log in to Amazon Vine only once per pass, so that we don't perform
multiple simultaneous logins back-to-back.  This should make it harder
for Amazon to detect our script, as well as making it more robust.

Also support image captchas for verified logins.  An image will open
in a browser window, and the user will be prompted to type in the word.
This will be repeated until the user gets it right.

Finally, display a proper message and abort if the userid or login are
unknown.