Ignore items that would exceed the $600 tax limit
Add a command-line option (-t) that tells the script to ignores
items that would put you over the $600 tax limit.

Amazon will send you and the IRS a form 1099 if you receive more
than $600 worth of items between July 1st and June 30th of the
following year.  You will then have to pay taxes on the value
of all of those items.

Technically, you're supposed to pay taxes even if you receive less
than $600 and don't get a form 1099, but no one does that.

The problem is that once you cross that $600 threshold, you will
suddenly have to pay around $200 in taxes, and you will probably
not be able to sell enough items to recoup that cost.  This means
that it's probably better to stay below $600, even if it means
not receiving any items you would like.