Use cookies from an existing web browser login
About six months ago, something changed on Amazon.com that would prevent
Mechanize from logging in.  Normal web browsers, even with Javascript
disabled, can log in fine.  But when this script tries to log in, Amazon
responds with an "invalid email" error message.

My theory is that Amazon has changed the cookies it sends, and Mechanize is
uniquely unable to process those cookies.

Fortunately, there is a work-around.  If you log into Amazon with a normal
web browser, and then copy the cookies from the browser, this script can
then log in.

This script now depends on the browsercookie package at

	https://pypi.python.org/pypi/browsercookie/

To load cookies from either Firefox or Chrome.  If you use Chrome, you'll be
prompted with a keychain security prompt.

Steps:

1) Load the browser
2) Make sure cookies are fully enabled.
3) Log into Amazon.
4) Quit the browser (this will ensure the cookies are saved to disk)
5) Launch the script.  Specify the --browser option if you used Chrome