Import third-party modules in a better way
Move the code that loads third-party modules into its own function,
and call that function after parsing command-line parameters.  Some
modules take a while to initialize, and we don't want that if the
user just asked for the help text.

Also display the version number of each module that's loaded, if
available.